refactor(vault): extract vault-search and vault-stats query modules
Move searchInContents/searchNotes plus the scoring/snippet helpers and
search constants into src/lib/vault-search.ts (depends on note-reads, not
vault-fs directly, since searchNotes reads via readNote/listNotes). Move
getNoteStats/listAttachments/getAttachmentStats plus the private
walkVaultExcluding walker into src/lib/vault-stats.ts (vault-fs only).
vault.ts re-exports all moved symbols so the ./vault.js API is unchanged.
